+++ api/src/db/seeds/production/roles.js
@@ -1,0 +1,7 @@
+exports.seed = (knex) => {
+  return knex('roles')
+    .del()
+    .then(() => {
+      return knex('roles').insert([{ name: 'admin' }])
+    })
+}