+++ pages/campaign.js
@@ -1,14 +1,20 @@
 import React, { Component } from 'react'
-import '../styles/Campaigns.scss'
 import Link from 'next/link'
-import Router from 'next/router'
-import api from '../lib/utils/api'
+import dynamic from 'next/dynamic'
+import { connect } from 'unistore/react'
 import { distanceInWordsToNow } from 'date-fns'
-import CampaignMap from '../components/charts/CampaignMap'
+import Router from 'next/router'
+
+import api from '../lib/utils/api'
+import { actions } from '../lib/store'
 import UserTable from '../components/UserTable'
 import ReactMarkdown from 'react-markdown'
 import { formatDecimal } from '../lib/utils/format'
 import sumEdits from '../lib/utils/sum_edits'
+
+const CampaignMap = dynamic(() => import('../components/charts/CampaignMap'), {
+  ssr: false
+})
 
 const Blurb = ({
   users,
@@ -26,112 +32,103 @@ const Blurb = ({
   </h2>
 }
 
-class Campaign extends Component {
-  constructor () {
-    super()
-    this.state = {
-      records: {},
-      notFound: false
-    }
-  }
+export class Campaign extends Component {
+  render () {
+    const { records } = this.props
 
-  componentDidMount () {
-    const { match } = this.props
-    /**
-     * TODO:
-     * What to do if there is no match?
-     */
-    if (match) {
-      const { params: { name } } = match
-      api('get', `/api/campaigns/${name}`)
-        .then(res => {
-          // TODO error state
-          this.setState({
-            records: res.data.records,
-            match
-          })
-        }).catch((e) => {
-          console.log(e)
-          this.setState({ notFound: true })
-        })
+    if (!records || !records.tmData) {
+      return <div />
     }
-  }
 
-  renderRedirect () {
-    if (this.state.notFound) {
+    return (
+      <div className='Campaigns'>
+        <header className='header--internal--green header--page'>
+          <div className='row'>
+            <div className='section-sub--left'>
+              <h1 className='header--xlarge margin-top-sm'>{records.tmData.name}</h1>
+              <ul className='list--two-column clearfix'>
+                <li>
+                  <span className='list-label'>Project Number:</span>
+                  <span>#{records.tmData.tm_id}</span>
+                </li>
+                <li>
+                  <span className='list-label'>Last Update:</span>
+                  <span>{distanceInWordsToNow(records.tmData.updated_at)} ago.</span>
+                </li>
+              </ul>
+            </div>
+            <div className='section-sub--right'>
+              <Link href='/about'>
+                <a className='button'>Contribute</a>
+              </Link>
+            </div>
+          </div>
+        </header>
+        <section>
+          <div className='row'>
+            <div className='section-sub--left section-width-fifty-plus'>
+              <div className='text-body'><ReactMarkdown source={records.tmData.description} /></div>
+            </div>
+            <div className='section-sub--right section-width-fifty-minus'>
+              <div className='map-campaign-lg'>
+                <div className='campagin-body-header'>
+                  <ul className='list--horizontal'>
+                    <li className='list--inline'>
+                      <span className='descriptor-chart'>Complete</span>
+                      <span className='num--large'>{parseInt(records.tmData.done * 0.5, 10) + parseInt(records.tmData.validated, 10)}%</span>
+                    </li>
+                    <li className='list--inline'>
+                      <span className='descriptor-chart'>Participants</span>
+                      <span className='num--large'>{records.users.length}</span>
+                    </li>
+                    <li className='list--inline'>
+                      <span className='descriptor-chart'>Total Features Mapped</span>
+                      <span className='num--large'>
+                        {formatDecimal(sumEdits(records))}
+                      </span>
+                    </li>
+                  </ul>
+                </div>
+                <CampaignMap feature={JSON.parse(records.tmData.geometry)} interactive />
+              </div>
+            </div>
+          </div>
+        </section>
+        <section className='section--tertiary'>
+          <div className='row'>
+            <Blurb {...records} />
+            <UserTable users={records.users} />
+          </div>
+        </section>
+      </div>
+    )
+  }
+}
+
+const Page = connect(['authenticatedUser'], actions)(Campaign)
+
+Page.getInitialProps = async ({ server, req, res }) => {
+  const { id } = req.params
+
+  function redirect () {
+    if (typeof window === 'undefined') {
+      return res.redirect('/404')
+    } else {
       return Router.push('/404')
     }
   }
 
-  render () {
-    const { records, match } = this.state
-    if (match && records && records.tmData) {
-      return (
-        <div className='Campaigns'>
-          <header className='header--internal--green header--page'>
-            <div className='row'>
-              <div className='section-sub--left'>
-                <h1 className='header--xlarge margin-top-sm'>{records.tmData.name}</h1>
-                <ul className='list--two-column clearfix'>
-                  <li>
-                    <span className='list-label'>Project Number:</span>
-                    <span>#{records.tmData.tm_id}</span>
-                  </li>
-                  <li>
-                    <span className='list-label'>Last Update:</span>
-                    <span>{distanceInWordsToNow(records.tmData.updated_at)} ago.</span>
-                  </li>
-                </ul>
-              </div>
-              <div className='section-sub--right'>
-                <Link href='/about'>
-                  <a className='button'>Contribute</a>
-                </Link>
-              </div>
-            </div>
-          </header>
-          <section>
-            <div className='row'>
-              <div className='section-sub--left section-width-fifty-plus'>
-                <div className='text-body'><ReactMarkdown source={records.tmData.description} /></div>
-              </div>
-              <div className='section-sub--right section-width-fifty-minus'>
-                <div className='map-campaign-lg'>
-                  <div className='campagin-body-header'>
-                    <ul className='list--horizontal'>
-                      <li className='list--inline'>
-                        <span className='descriptor-chart'>Complete</span>
-                        <span className='num--large'>{parseInt(records.tmData.done * 0.5, 10) + parseInt(records.tmData.validated, 10)}%</span>
-                      </li>
-                      <li className='list--inline'>
-                        <span className='descriptor-chart'>Participants</span>
-                        <span className='num--large'>{records.users.length}</span>
-                      </li>
-                      <li className='list--inline'>
-                        <span className='descriptor-chart'>Total Features Mapped</span>
-                        <span className='num--large'>
-                          {formatDecimal(sumEdits(records))}
-                        </span>
-                      </li>
-                    </ul>
-                  </div>
-                  <CampaignMap feature={JSON.parse(records.tmData.geometry)} interactive />
-                </div>
-              </div>
-            </div>
-          </section>
-          <section className='section--tertiary'>
-            <div className='row'>
-              <Blurb {...records} />
-              <UserTable users={records.users} />
-            </div>
-          </section>
-        </div>
-      )
-    } else {
-      return <div>{this.renderRedirect()}</div>
+  try {
+    const { data } = await api('get', `/api/campaigns/${decodeURIComponent(id)}`)
+
+    if (!data.records || !data.records.tmData) {
+      return redirect()
     }
+
+    return data
+  } catch (e) {
+    redirect()
   }
 }
 
-export default Campaign
+export default Page