+++ pages/dashboard.js
@@ -5,7 +5,6 @@ import { connect } from 'unistore/react'
 import { actions } from '../lib/store'
 
 import BadgeInProgress from '../components/BadgeInProgress'
-import '../styles/Dashboard.scss'
 
 import NotLoggedIn from '../components/NotLoggedIn'
 
@@ -23,7 +22,6 @@ class Dashboard extends Component {
 
   componentDidMount () {
     this.props.getAuthenticatedUser()
-    // this.props.getProjects()
   }
 
   componentDidUpdate () {
@@ -127,6 +125,7 @@ class Dashboard extends Component {
 
     // We are logged in and should have a profile
     const osmUser = osm._xml2json.user
+    const badgeNums = account.badges ? Object.keys(account.badges.earnedBadges).length : 0
 
     return (
       <div className='dashboard'>
@@ -146,7 +145,7 @@ class Dashboard extends Component {
                   </li>
                   <li className='list--inline'>
                     <span className='descriptor-chart'>Badges</span>
-                    <span className='num--large'>{Object.keys(account.badges.earnedBadges).length}</span>
+                    <span className='num--large'>{badgeNums}</span>
                   </li>
                   <li className='list--inline'>
                     <span className='descriptor-chart'>Edits</span>
@@ -160,7 +159,7 @@ class Dashboard extends Component {
           {/* <UserExtentMap extent={user.records.extent_uri} uid={osmUser['@']['id']} /> */}
         </header>
 
-        {this.renderUpcomingBadges(account.badges.unearnedBadges)}
+        {account.badges ? this.renderUpcomingBadges(account.badges.unearnedBadges) : ''}
         {/* this.renderProjects() */}
       </div>
     )