+++ pages/edit-user.js
@@ -8,9 +8,6 @@ import { actions } from '../lib/store'
 
 import NotLoggedIn from '../components/NotLoggedIn'
 
-import '../styles/Users.scss'
-import 'react-select/dist/react-select.css'
-
 class UserEdit extends Component {
   constructor (props) {
     super(props)
@@ -18,35 +15,24 @@ class UserEdit extends Component {
   }
 
   updateUser (data) {
-    const { user: { id } } = this.props
+    const { id } = this.props
 
     api('put', `/api/users/${id}`, data)
       .then(res => {
         this.setState({ saved: true })
+        this.props.setNotification({ type: 'info', message: '✓ Saved' })
       })
       .catch(err => {
-        // TODO: show error
-        console.log('err', err)
+        console.log(err)
+        this.props.setNotification({ type: 'error', message: 'Could not update user' })
       })
   }
 
-  onCountryChange = (country) => {
+  onCountryChange (country) {
     const { currentCountry } = this.state
     if (!country || country.value === currentCountry) return
     this.setState({ currentCountry: country.value })
     this.updateUser({ country: country.value })
-  }
-
-  renderSaved () {
-    if (this.state.saved) {
-      setTimeout(() => {
-        this.setState({ saved: false })
-      }, 1000)
-
-      return (
-        <p style={{ color: '#4FCA9E' }}>✓ Saved</p>
-      )
-    }
   }
 
   render () {
@@ -72,9 +58,8 @@ class UserEdit extends Component {
               <Select
                 options={countries}
                 value={country}
-                onChange={this.onCountryChange}
+                onChange={(country) => this.onCountryChange(country)}
               />
-              {this.renderSaved()}
             </div>
           </div>
         </section>