+++ pages/users.js
@@ -1,6 +1,4 @@
 import React, { Component } from 'react'
-import '../styles/Users.scss'
-import '../styles/Campaigns.scss'
 import { equals, pick } from 'ramda'
 import queryString from 'query-string'
 import Pagination from 'react-js-pagination'
@@ -50,8 +48,11 @@ export class Users extends Component {
         this.setState(Object.assign({ records: res.data, apiStatus: 'SUCCESS' }))
       })
       .catch(err => {
-        console.log('err', err)
-        this.setState({ apiStatus: 'ERROR' })
+        console.log(err)
+        if (this.state.apiStatus !== 'ERROR') {
+          this.setState({ apiStatus: 'ERROR' })
+          this.props.setNotification({ type: 'error', message: 'Could not retrieve user stats' })
+        }
       })
   }
 
@@ -121,7 +122,7 @@ export class Users extends Component {
     } = this.props.users
 
     return (
-      <div className='Users'>)
+      <div className='Users'>
         <AllUsersHeader countries={countries} users={total} edits={editTotal} />
         <section>
           <div className='row'>