+++ tests/containers.test.js
@@ -4,8 +4,8 @@ import { Provider } from 'unistore/react'
 
 // import {Campaigns, Campaign, Home, User, Users, About} from '../containers';
 import { Home } from '../pages/index'
-import Campaigns from '../pages/campaigns'
-import Campaign from '../pages/campaign'
+import { Campaigns } from '../pages/campaigns'
+import { Campaign } from '../pages/campaign'
 import User from '../pages/user'
 import { Users } from '../pages/users'
 import About from '../pages/about'
@@ -18,7 +18,19 @@ function mockAction () {
 
 it('Campaigns renders without crashing', () => {
   const div = document.createElement('div')
-  const mockProps = {}
+  const mockProps = {
+    campaigns: {
+      searchText: '',
+      compl_min: 0,
+      compl_max: 100,
+      page: 1,
+      apiStatus: 'LOADING',
+      records: {}
+    },
+    handleCampaignsSearch: mockAction,
+    handleCampaignsFilterChange: mockAction,
+    handleCampaignsPageChange: mockAction
+  }
 
   const page = (
     <Provider store={store}>