+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 - Ability to login using OSM (using passport in the backend)
-- Initial dashboard page
+- Dashboard page that allows users to see their personal stats
 - Allow badge creation, deletion, and updates through API routes
 - cli tool for debugging and administration in development
 - User roles
@@ -15,10 +15,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - role validation helpers
   - API endpoints
   - UI for managing the roles of users
-- Initial admin UI
+- Admin UI that allows admin users to manage users, roles, and badges
 - Mock passport strategy for testing routes that require authentication
+- Support for badges based on the time edits are made and the hashtags they're made with
+- Hashtag-based badges and expired date-specific badges are excluded from "In Progress" display
 
 ### Changed
+- Migrated to [next.js](https://github.com/zeit/next.js/)
+  - Instead of deploying the api and frontend separately, the two codebases are 
+    merged into one with server-rendering provided by next.js
+  - For changes to development process see the [updated README.md](README.md)
 - Move from Sqlite to Postgres
 - Drop some columns from the user schema
 - Move tests and fixtures for API into a single place
@@ -26,6 +32,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Move badge details to the database
 - Revise user model to be the same format as the roles model
 - Move existing `api/users` endpoint to `api/users/stats` and use `api/users` for a list of users without stats
+- Consecutive and total days mapped are calculated with dates formatted YYYY-MM-DD
 
 ## [v0.2.4] - 2018-10-23
 ### Added